Rename generators references to validators
