+++ services/blockchain-indexer/shared/database/schema/validators.js
